Split settings into smaller, more understandable chunks
