+++ api/Pipfile
@@ -54,6 +54,7 @@ redlock-py = "~=1.0"
 requests-oauthlib = "~=1.3"
 sentry-sdk = "~=1.19"
 wsgi-basic-auth = "~=1.1"
+django-split-settings = "*"
 
 [requires]
 python_version = "3.11"
