+++ api/conf/settings/openverse.py
@@ -1,0 +1,6 @@
+from conf.settings.base import INSTALLED_APPS
+
+
+INSTALLED_APPS += [
+    "api",
+]
